+++ Pandora/Core.hs
@@ -1,5 +1,4 @@
 module Pandora.Core (module Exports) where
 
-import Pandora.Core.Transformation as Exports
 import Pandora.Core.Morphism as Exports
 import Pandora.Core.Functor as Exports